Revert "Add hax for using buried nodes in EDIF."
This reverts commit 0311bc1087995699dbce599325ffbc687b7bb640.

It was stupid and insulting of me to think this doesn't work. In fact
all that's necessary is to specify the name of the buried node
(usually the FF cell) in the pin assignment list (rather than its
output net) for it to be placed appropriately. This should have been
obvious from the way it's specified in CUPL and from CUPL docs.