Reorganize the database a bit. NFC.
It looks like most of the MC config fuses (except for IO ones) are
intended to be 1 by default; which makes sense because that's
the erased flash state, but even then the fitter sets many of them
to 0 in unused macrocells. It could be that the fitter is trying
to conserve power, or just that there is no "default" polarity.

Regardless it looks a bit nicer now in the html.