Add ATF1508 fuse ranges and pinouts, disabled for now.
ATF1508 does not have any (remaining in production) packages with
every macrocell bonded out, which is a problem because about half
of the fuzzers use a constraint on an output pin, and there appears
to be no way to constrain an output to an internal macrocell (even
though those buffers are likely still accessible via JTAG).

A lot of work will be necessary if ATF1508 is to be ever supported.