Provide an explicit map from special functions to macrocells.
This is useful for determining which macrocells become unusable as IO
when the special function is activated. It also removes a few warts
from fuzzers, such as `gclk3_pad` and macrocell lists in the global
configuration fuzzer.

Also replace `power_down_X` with `pdX_pin_func` while we're at it.