Optionally cache toolchain runs.
It's pretty dumb that I *first* finished essentially all of
the fuzzers (frequently waiting a long time for chipdb to bootstrap)
and *then* bothered to make bootstrapping faster, but here it is,
toolchain caching.

The main reason this is useful is that I'd like to change
the database schema significantly prior to a proper public release,
and that involves a lot of bootstrapping without ever really asking
the fitter any new questions.

With a cold cache, a bootstrap:
  * takes 35 min on my laptop;
  * populates the cache with 231 MB in 32477 files.

With a warm cache, a bootstrap takes less than 30 sec. Pretty nice!